test_requirements.txt
pre-commit==4.0.1
pylint==3.3.3
pytest-homeassistant-custom-component==0.13.203
ruff==0.9.2
pytest==8.3.4
aioresponses==0.7.8